requirements.txt
scipy>=0.19.0
numpy>=1.12.1
pandas>=0.19.2
pyyaml==5.4.1
statsmodels
tensorflow-gpu==1.15.0
tables
lxml